requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-split-settings==1.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
frozenlist==1.5.0
# google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
hiredis==3.1.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
propcache==0.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyparsing==3.2.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
split-settings==1.0.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
